add listusers + fix register

+++ frontend/pages/listusers.php
@@ -2,14 +2,22 @@
 require("./auth/controller/AuthController.class.php");
 $users = new AuthController($_GET);
 $userData = $users->listUserTable();
-$users->removeUser();
+
+if (!empty($_GET['deleteid']) && isset($_GET['deleteid'])) {
+    $users->removeUser();
+} else if (!empty($_GET['editid']) && isset($_GET['editid'])) {
+    if (!empty($_POST) && isset($_POST)) {
+        $users->editUser($_POST);
+    }
+}
+
 ?>
 <div class="title">
     <h3>Lista utenti</h3>
 </div>
 <div class="row">
-    <div class="col-md-12">
-        <table class="table">
+    <div class="col-md-12" style="overflow-x: auto;">
+        <table class="table responsive">
             <thead>
                 <tr>
                 <?php foreach ($userData[0] as $header) { ?>
@@ -18,27 +26,48 @@ $users->removeUser();
                 </tr>
             </thead>
             <tbody>
-                <?php foreach (array_slice($userData, 1) as $user) {?>
+                <?php foreach (array_slice($userData, 1) as $user) {
+                if (!empty($_GET['editid']) && isset($_GET['editid']) && $_GET['editid'] == $user[4]) {
+                    echo '<form action="" method="post" id="edituser">';
+                }    
+                ?>
                 <tr>
-                    <?php foreach ($user as $key => $value) { 
+                    <?php foreach ($user as $key => $value) {
+                        $name = $key."-".$user[4];
+                        //apro la cella
                         echo '<td>';
-                        switch ($key) {
-                            case 3:
-                                $checked = $value ? 'checked' : '';    
-                                echo '<input class="form-check-input" type="checkbox" name="admin" '.$checked.' disabled>';
-                                break;
-                            case 4:
+                        //attribuisco il valore per ogni cella
+                        if ($key === 0 || $key === 1 || $key === 2) {
+                            if (!empty($_GET['editid']) && isset($_GET['editid']) && $_GET['editid'] == $user[4]) {
+                                echo '<input type="text" value="' . $value . '" name="' . $key . '">';
+                            } else {
+                                echo $value;
+                            }
+                        } elseif ($key === 3) {
+                            $checked = $value ? 'checked' : '';
+                            $disabled = !empty($_GET['editid']) && isset($_GET['editid']) && $_GET['editid'] == $user[4] ?: 'disabled';
+                            echo '<input class="form-check-input" type="checkbox" name="' . $key . '" ' . $checked . ' ' . $disabled . '>';
+                        } elseif ($key === 4) {
+                            if (!empty($_GET['editid']) && isset($_GET['editid']) && $_GET['editid'] == $user[4]) {
+                                echo '<a href="#" onclick="document.getElementById(\'edituser\').submit();">Conferma</a> - ' .
+                                    '<a href="'.refreshPageWOmsg().'">Chiudi</a>';
+                            } else {
                                 echo '<a href="index.php?page=listusers&editid=' . $value . '">Modifica</a> - ' .
                                     '<a href="index.php?page=listusers&deleteid=' . $value . '">Cancella</a>';
-                                break;
-                            default:
-                                echo $value;
+                            }
+                        } else {
+                            echo $value;
                         }
                         echo '</td>';
-                        }
+                        
+                    }
                     ?>
                 </tr>
-                <?php } ?>
+                <?php 
+                if (!empty($_GET['editid']) && isset($_GET['editid']) && $_GET['editid'] == $user[4]) {
+                    echo '</form>';
+                }
+            } ?>
             </tbody>
         </table>
     </div>
